Remove the restriction on schema type.
The validation of the schema itself restricted the type of the root
object of the schema to an 'object' or an 'array'. This means that
the following schema would be rejected:
  {
    type: 'null',
    description: 'Nothing is returned.'
  }

This fix removes this limitation.

Also updated the test suite following the change in requirement
for the root type of a schema.

(Also added a semi-colon missing at the end of a line...)