Update date-time format acceptance criteria.
According to ISO8601 spec (http://en.wikipedia.org/wiki/ISO_8601#Times),
the time can include fractions.

The JSON schema spec ( draft 3 ) recommends but does not force the
following ISO8601 format: YYYY-MM-DDThh:mm:ssZ.

Since a javascript Date object includes the milliseconds when converted
into a string with method toISOString(), I thought easier to accept the
string with the milliseconds in the JSON Schema. The code was modified
to accept but not force the presence of a 'dot' and then an unspecified
number of digits representing the fraction of a second.

I ran the test suite and nothing failed.